requirements.txt
matplotlib
scikit-image
Pillow
numpy
opencv-python
plotly